requirements.txt
aws-cdk-lib==2.69.0
constructs>=10.0.0,<11.0.0
dynaconf==2.2.3